Fixed Bitcoin/Solana POMs
